Rename Spatial to SpatialLayout and Scrolling to ScrollingLayout
