+++ src/window_manager/tests/mod.rs
@@ -2,7 +2,7 @@ use crate::window_manager::window_manager;
 
 mod navigation_tests;
 mod placement_tests;
-mod scrolling_tests;
-mod spatial_tests;
+mod scrolling_layout_tests;
+mod spatial_layout_tests;
 mod test_support;
 mod window_manager_tests;
